Split up code into smaller component files
