+++ src/plugins/discordColorways/components/themePreview.tsx
@@ -1,0 +1,59 @@
+/*
+ * Vencord, a Discord client mod
+ * Copyright (c) 2023 Vendicated and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+import { Forms, Text, useState } from "@webpack/common";
+
+import { HexToHSL } from "../utils";
+
+export function ThemePreviewCategory({ accent, primary, secondary, tertiary, className, isCollapsed, previewCSS }: { accent: string, primary: string, secondary: string, tertiary: string, className?: string, isCollapsed: boolean, previewCSS?: string; }) {
+    const [collapsed, setCollapsed] = useState<boolean>(isCollapsed);
+    return (<div className={`${collapsed === true ? "colorwaysPreview colorwaysPreview-collapsed" : "colorwaysPreview"} ${className}`}>
+        <div className="colorwaysCreator-settingItm colorwaysCreator-settingHeader" onClick={() => collapsed === true ? setCollapsed(false) : setCollapsed(true)}><Forms.FormTitle style={{ marginBottom: 0 }}>Preview</Forms.FormTitle><svg className="expand-3Nh1P5 transition-30IQBn directionDown-2w0MZz" width="24" height="24" viewBox="0 0 24 24" aria-hidden="true" role="img"><path fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" d="M7 10L12 15 17 10" aria-hidden="true"></path></svg></div>
+        <style>{previewCSS}</style>
+        <ThemePreview accent={accent} primary={primary} secondary={secondary} tertiary={tertiary} />
+    </div>);
+}
+
+export function ThemePreview({ accent, primary, secondary, tertiary, previewCSS }: { accent: string, primary: string, secondary: string, tertiary: string, previewCSS?: string; }) {
+    return (<div className="colorwaysPreview-container">
+        <style>{previewCSS}</style>
+        <div className="colorwaysPreview-wrapper" style={{ backgroundColor: tertiary }}>
+            <div className="colorwaysPreview-titlebar"></div>
+            <div className="colorwaysPreview-body">
+                <div className="colorwayPreview-guilds">
+                    <div className="colorwayPreview-guild">
+                        <div className="colorwayPreview-guildItem" style={{ backgroundColor: primary }} onMouseEnter={e => { e.currentTarget.style.backgroundColor = accent; }} onMouseLeave={e => { e.currentTarget.style.backgroundColor = primary; }} onClick={e => {
+                            e.currentTarget.parentElement?.parentElement?.parentElement?.parentElement?.requestFullscreen();
+                        }}>
+                            <svg className="controlIcon-10O-4h" aria-hidden="true" role="img" width="24" height="24" viewBox="0 0 24 24"><path fill="currentColor" d="M19,3H14V5h5v5h2V5A2,2,0,0,0,19,3Z"></path><path fill="currentColor" d="M19,19H14v2h5a2,2,0,0,0,2-2V14H19Z"></path><path fill="currentColor" d="M3,5v5H5V5h5V3H5A2,2,0,0,0,3,5Z"></path><path fill="currentColor" d="M5,14H3v5a2,2,0,0,0,2,2h5V19H5Z"></path></svg>
+                        </div>
+                    </div>
+                    <div className="colorwayPreview-guild">
+                        <div className="colorwayPreview-guildSeparator" style={{ backgroundColor: primary }}></div>
+                    </div>
+                    <div className="colorwayPreview-guild">
+                        <div className="colorwayPreview-guildItem" style={{ backgroundColor: primary }} onMouseEnter={e => { e.currentTarget.style.backgroundColor = accent; }} onMouseLeave={e => { e.currentTarget.style.backgroundColor = primary; }}></div>
+                    </div>
+                    <div className="colorwayPreview-guild">
+                        <div className="colorwayPreview-guildItem" style={{ backgroundColor: primary }} onMouseEnter={e => { e.currentTarget.style.backgroundColor = accent; }} onMouseLeave={e => { e.currentTarget.style.backgroundColor = primary; }}></div>
+                    </div>
+                </div>
+                <div className="colorwayPreview-channels" style={{ backgroundColor: secondary }}>
+                    <div className="colorwayPreview-userArea" style={{ backgroundColor: "hsl(" + HexToHSL(secondary)[0] + " " + HexToHSL(secondary)[1] + "% " + Math.max(HexToHSL(secondary)[2] - 3.6, 0) + "%)" }}></div>
+                    <div className="colorwayPreview-filler"></div>
+                    <div className="colorwayPreview-topShadow" style={{ "--primary-900-hsl": `${HexToHSL(tertiary)[0]} ${HexToHSL(tertiary)[1]}% ${Math.max(HexToHSL(tertiary)[2] - (3.6 * 6), 0)}%`, "--primary-500-hsl": `${HexToHSL(primary)[0]} ${HexToHSL(primary)[1]}% ${Math.min(HexToHSL(primary)[2] + (3.6 * 3), 100)}%` } as React.CSSProperties}>
+                        <Text tag="div" variant="text-md/semibold" lineClamp={1} selectable={false}>Preview</Text>
+                    </div>
+                </div>
+                <div className="colorwayPreview-chat" style={{ backgroundColor: primary }}>
+                    <div className="colorwayPreview-chatBox" style={{ backgroundColor: "hsl(" + HexToHSL(primary)[0] + " " + HexToHSL(primary)[1] + "% " + Math.min(HexToHSL(primary)[2] + 3.6, 100) + "%)" }}></div>
+                    <div className="colorwayPreview-filler"></div>
+                    <div className="colorwayPreview-topShadow" style={{ "--primary-900-hsl": `${HexToHSL(tertiary)[0]} ${HexToHSL(tertiary)[1]}% ${Math.max(HexToHSL(tertiary)[2] - (3.6 * 6), 0)}%` } as React.CSSProperties}></div>
+                </div>
+            </div>
+        </div>
+    </div>);
+}
