+++ src/plugins/discordColorways/css.ts
@@ -1,0 +1,393 @@
+/*
+ * Vencord, a Discord client mod
+ * Copyright (c) 2023 Vendicated and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+import { HexToHSL } from "./utils";
+
+export const colorVariables: string[] = [
+    "brand-100",
+    "brand-130",
+    "brand-160",
+    "brand-200",
+    "brand-230",
+    "brand-260",
+    "brand-300",
+    "brand-330",
+    "brand-345",
+    "brand-360",
+    "brand-400",
+    "brand-430",
+    "brand-460",
+    "brand-500",
+    "brand-530",
+    "brand-560",
+    "brand-600",
+    "brand-630",
+    "brand-660",
+    "brand-700",
+    "brand-730",
+    "brand-760",
+    "brand-800",
+    "brand-830",
+    "brand-860",
+    "brand-900",
+    "primary-900",
+    "primary-860",
+    "primary-830",
+    "primary-800",
+    "primary-760",
+    "primary-730",
+    "primary-700",
+    "primary-660",
+    "primary-645",
+    "primary-630",
+    "primary-600",
+    "primary-560",
+    "primary-530",
+    "primary-500",
+    "primary-460",
+    "primary-430",
+    "primary-400",
+    "primary-360",
+    "primary-330",
+    "primary-300",
+    "primary-260",
+    "primary-230",
+    "primary-200",
+    "primary-160",
+    "primary-130",
+    "primary-100",
+    "white-900",
+    "white-860",
+    "white-830",
+    "white-800",
+    "white-760",
+    "white-730",
+    "white-700",
+    "white-660",
+    "white-630",
+    "white-600",
+    "white-560",
+    "white-530",
+    "white-500",
+    "white-460",
+    "white-430",
+    "white-400",
+    "white-360",
+    "white-330",
+    "white-300",
+    "white-260",
+    "white-230",
+    "white-200",
+    "white-160",
+    "white-130",
+    "white-100",
+    "teal-900",
+    "teal-860",
+    "teal-830",
+    "teal-800",
+    "teal-760",
+    "teal-730",
+    "teal-700",
+    "teal-660",
+    "teal-630",
+    "teal-600",
+    "teal-560",
+    "teal-530",
+    "teal-500",
+    "teal-460",
+    "teal-430",
+    "teal-400",
+    "teal-360",
+    "teal-330",
+    "teal-300",
+    "teal-260",
+    "teal-230",
+    "teal-200",
+    "teal-160",
+    "teal-130",
+    "teal-100",
+    "black-900",
+    "black-860",
+    "black-830",
+    "black-800",
+    "black-760",
+    "black-730",
+    "black-700",
+    "black-660",
+    "black-630",
+    "black-600",
+    "black-560",
+    "black-530",
+    "black-500",
+    "black-460",
+    "black-430",
+    "black-400",
+    "black-360",
+    "black-330",
+    "black-300",
+    "black-260",
+    "black-230",
+    "black-200",
+    "black-160",
+    "black-130",
+    "black-100",
+    "red-900",
+    "red-860",
+    "red-830",
+    "red-800",
+    "red-760",
+    "red-730",
+    "red-700",
+    "red-660",
+    "red-630",
+    "red-600",
+    "red-560",
+    "red-530",
+    "red-500",
+    "red-460",
+    "red-430",
+    "red-400",
+    "red-360",
+    "red-330",
+    "red-300",
+    "red-260",
+    "red-230",
+    "red-200",
+    "red-160",
+    "red-130",
+    "red-100",
+    "yellow-900",
+    "yellow-860",
+    "yellow-830",
+    "yellow-800",
+    "yellow-760",
+    "yellow-730",
+    "yellow-700",
+    "yellow-660",
+    "yellow-630",
+    "yellow-600",
+    "yellow-560",
+    "yellow-530",
+    "yellow-500",
+    "yellow-460",
+    "yellow-430",
+    "yellow-400",
+    "yellow-360",
+    "yellow-330",
+    "yellow-300",
+    "yellow-260",
+    "yellow-230",
+    "yellow-200",
+    "yellow-160",
+    "yellow-130",
+    "yellow-100",
+    "green-900",
+    "green-860",
+    "green-830",
+    "green-800",
+    "green-760",
+    "green-730",
+    "green-700",
+    "green-660",
+    "green-630",
+    "green-600",
+    "green-560",
+    "green-530",
+    "green-500",
+    "green-460",
+    "green-430",
+    "green-400",
+    "green-360",
+    "green-330",
+    "green-300",
+    "green-260",
+    "green-230",
+    "green-200",
+    "green-160",
+    "green-130",
+    "green-100",
+];
+
+
+export function generateCss(primaryColor: string, secondaryColor: string, tertiaryColor: string, accentColor: string, tintedText: boolean) {
+    return `/*Automatically Generated - Colorway Creator V${(Vencord.Plugins.plugins.DiscordColorways as any).creatorVersion}*/
+:root {
+    --brand-100-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 13)), 100)}%;
+    --brand-130-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 12)), 100)}%;
+    --brand-160-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 11)), 100)}%;
+    --brand-200-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 10)), 100)}%;
+    --brand-230-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 9)), 100)}%;
+    --brand-260-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 8)), 100)}%;
+    --brand-300-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 7)), 100)}%;
+    --brand-330-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 6)), 100)}%;
+    --brand-345-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 5)), 100)}%;
+    --brand-360-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 4)), 100)}%;
+    --brand-400-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 3)), 100)}%;
+    --brand-430-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + (3.6 * 2)), 100)}%;
+    --brand-460-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.min(Math.round(HexToHSL("#" + accentColor)[2] + 3.6), 100)}%;
+    --brand-500-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${HexToHSL("#" + accentColor)[2]}%;
+    --brand-530-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - 3.6), 0)}%;
+    --brand-560-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 2)), 0)}%;
+    --brand-600-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 3)), 0)}%;
+    --brand-630-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 4)), 0)}%;
+    --brand-660-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 5)), 0)}%;
+    --brand-700-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 6)), 0)}%;
+    --brand-730-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 7)), 0)}%;
+    --brand-760-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 8)), 0)}%;
+    --brand-800-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 9)), 0)}%;
+    --brand-830-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 10)), 0)}%;
+    --brand-860-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 11)), 0)}%;
+    --brand-900-hsl: ${HexToHSL("#" + accentColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + accentColor)[1]}%) ${Math.max(Math.round(HexToHSL("#" + accentColor)[2] - (3.6 * 12)), 0)}%;
+    --primary-800-hsl: ${HexToHSL("#" + tertiaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + tertiaryColor)[1]}%) ${Math.min(HexToHSL("#" + tertiaryColor)[2] + (3.6 * 2), 100)}%;
+    --primary-730-hsl: ${HexToHSL("#" + tertiaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + tertiaryColor)[1]}%) ${Math.min(HexToHSL("#" + tertiaryColor)[2] + 3.6, 100)}%;
+    --primary-700-hsl: ${HexToHSL("#" + tertiaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + tertiaryColor)[1]}%) ${HexToHSL("#" + tertiaryColor)[2]}%;
+    --primary-660-hsl: ${HexToHSL("#" + secondaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + secondaryColor)[1]}%) ${Math.min(HexToHSL("#" + secondaryColor)[2] + 3.6, 100)}%;
+    --primary-645-hsl: ${HexToHSL("#" + primaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + primaryColor)[1]}%) ${Math.max(HexToHSL("#" + primaryColor)[2] - 5, 0)}%;
+    --primary-630-hsl: ${HexToHSL("#" + secondaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + secondaryColor)[1]}%) ${HexToHSL("#" + secondaryColor)[2]}%;
+    --primary-600-hsl: ${HexToHSL("#" + primaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + primaryColor)[1]}%) ${HexToHSL("#" + primaryColor)[2]}%;
+    --primary-560-hsl: ${HexToHSL("#" + primaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + primaryColor)[1]}%) ${Math.min(HexToHSL("#" + primaryColor)[2] + 3.6, 100)}%;
+    --primary-530-hsl: ${HexToHSL("#" + primaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + primaryColor)[1]}%) ${Math.min(HexToHSL("#" + primaryColor)[2] + (3.6 * 2), 100)}%;
+    --primary-500-hsl: ${HexToHSL("#" + primaryColor)[0]} calc(var(--saturation-factor, 1)*${HexToHSL("#" + primaryColor)[1]}%) ${Math.min(HexToHSL("#" + primaryColor)[2] + (3.6 * 3), 100)}%;${tintedText ? `\n    --primary-460-hsl: 0 calc(var(--saturation-factor, 1)*0%) 50%;
+    --primary-430: ${HexToHSL("#" + secondaryColor)[0] === 0 ? "gray" : ((HexToHSL("#" + secondaryColor)[2] < 80) ? "hsl(" + HexToHSL("#" + secondaryColor)[0] + ", calc(var(--saturation-factor, 1)*100%), 90%)" : "hsl(" + HexToHSL("#" + secondaryColor)[0] + ", calc(var(--saturation-factor, 1)*100%), 20%)")};
+    --primary-400: ${HexToHSL("#" + secondaryColor)[0] === 0 ? "gray" : ((HexToHSL("#" + secondaryColor)[2] < 80) ? "hsl(" + HexToHSL("#" + secondaryColor)[0] + ", calc(var(--saturation-factor, 1)*100%), 90%)" : "hsl(" + HexToHSL("#" + secondaryColor)[0] + ", calc(var(--saturation-factor, 1)*100%), 20%)")};
+    --primary-360: ${HexToHSL("#" + secondaryColor)[0] === 0 ? "gray" : ((HexToHSL("#" + secondaryColor)[2] < 80) ? "hsl(" + HexToHSL("#" + secondaryColor)[0] + ", calc(var(--saturation-factor, 1)*100%), 90%)" : "hsl(" + HexToHSL("#" + secondaryColor)[0] + ", calc(var(--saturation-factor, 1)*100%), 20%)")};` : ""}
+}${(Math.round(HexToHSL("#" + primaryColor)[2]) > 80) ? `\n\n/*Primary*/
+.theme-dark .container-2cd8Mz,
+.theme-dark .body-16rSsp,
+.theme-dark .toolbar-3_r2xA,
+.theme-dark .container-89zvna,
+.theme-dark .messageContent-2t3eCI,
+.theme-dark .attachButtonPlus-3IYelE,
+.theme-dark .username-h_Y3Us:not([style]),
+.theme-dark .children-3xh0VB,
+.theme-dark .buttonContainer-1502pf,
+.theme-dark .listItem-3SmSlK,
+.theme-dark .body-16rSsp .caret-1le2LN,
+.theme-dark .body-16rSsp .titleWrapper-24Kyzc > h1,
+.theme-dark .body-16rSsp .icon-2xnN2Y {
+    --white-500: black !important;
+    --interactive-normal: black !important;
+    --text-normal: black !important;
+    --text-muted: black !important;
+    --header-primary: black !important;
+    --header-secondary: black !important;
+}
+
+.theme-dark .contentRegionScroller-2_GT_N :not(.mtk1,.mtk2,.mtk3,.mtk4,.mtk5,.mtk6,.mtk7,.mtk8,.mtk9,.monaco-editor .line-numbers) {
+    --white-500: black !important;
+}
+
+.theme-dark .container-1um7CU,
+.theme-dark .container-2IKOsH,
+.theme-dark .header-3xB4vB {
+    background: transparent;
+}
+
+.theme-dark .container-ZMc96U {
+    --channel-icon: black;
+}
+
+.theme-dark .callContainer-HtHELf {
+    --white-500: ${(HexToHSL("#" + tertiaryColor)[2] > 80) ? "black" : "white"} !important;
+}
+
+.theme-dark .channelTextArea-1FufC0 {
+    --text-normal: ${(HexToHSL("#" + primaryColor)[2] + 3.6 > 80) ? "black" : "white"};
+}
+
+.theme-dark .placeholder-1rCBhr {
+    --channel-text-area-placeholder: ${(HexToHSL("#" + primaryColor)[2] + 3.6 > 80) ? "black" : "white"};
+    opacity: .6;
+}
+
+.theme-dark .colorwaySelectorIcon {
+    background-color: black;
+}
+
+.theme-dark .root-1CAIjD > .header-1ffhsl > h1 {
+    color: black;
+}
+/*End Primary*/`: ""}${(HexToHSL("#" + secondaryColor)[2] > 80) ? `\n\n/*Secondary*/
+.theme-dark .wrapper-2RrXDg *,
+.theme-dark .sidebar-1tnWFu *:not(.hasBanner-2IrYih *),
+.theme-dark .members-3WRCEx *:not([style]),
+.theme-dark .sidebarRegionScroller-FXiQOh *,
+.theme-dark .header-1XpmZs,
+.theme-dark .lookFilled-1H2Jvj.colorPrimary-2-Lusz {
+    --white-500: black !important;
+    --channels-default: black !important;
+    --channel-icon: black !important;
+    --interactive-normal: var(--white-500);
+    --interactive-hover: var(--white-500);
+    --interactive-active: var(--white-500);
+}
+
+.theme-dark .channelRow-4X_3fi {
+    background-color: var(--background-secondary);
+}
+
+.theme-dark .channelRow-4X_3fi * {
+    --channel-icon: black;
+}
+
+.theme-dark #app-mount .activity-2EQDZv {
+    --channels-default: var(--white-500) !important;
+}
+
+.theme-dark .nameTag-sc-gpq {
+    --header-primary: black !important;
+    --header-secondary: ${HexToHSL("#" + secondaryColor)[0] === 0 ? "gray" : ((HexToHSL("#" + secondaryColor)[2] < 80) ? "hsl(" + HexToHSL("#" + secondaryColor)[0] + ", calc(var(--saturation-factor, 1)*100%), 90%)" : "hsl(" + HexToHSL("#" + secondaryColor)[0] + ", calc(var(--saturation-factor, 1)*100%), 20%)")} !important;
+}
+
+.theme-dark .bannerVisible-Vkyg1I .headerContent-2SNbie {
+    color: #fff;
+}
+
+.theme-dark .embedFull-1HGV2S {
+    --text-normal: black;
+}
+/*End Secondary*/`: ""}${HexToHSL("#" + tertiaryColor)[2] > 80 ? `\n\n/*Tertiary*/
+.theme-dark .winButton-3UMjdg,
+.theme-dark .searchBar-2aylmZ *,
+.theme-dark .wordmarkWindows-2dq6rw,
+.theme-dark .searchBar-jGtisZ *,
+.theme-dark .searchBarComponent-3N7dCG {
+    --white-500: black !important;
+}
+
+.theme-dark [style="background-color: var(--background-secondary);"] {
+    color: ${HexToHSL("#" + secondaryColor)[2] > 80 ? "black" : "white"};
+}
+
+.theme-dark .popout-TdhJ6Z > *,
+.theme-dark .colorwayHeaderTitle {
+    --interactive-normal: black !important;
+    --header-secondary: black !important;
+}
+
+.theme-dark .tooltip-33Jwqe {
+    --text-normal: black !important;
+}
+/*End Tertiary*/`: ""}${HexToHSL("#" + accentColor)[2] > 80 ? `\n\n/*Accent*/
+.selected-2r1Hvo *,
+.selected-1Drb7Z *,
+#app-mount .lookFilled-1H2Jvj.colorBrand-2M3O3N:not(.buttonColor-3bP3fX),
+.colorDefault-2_rLdz.focused-3LIdPu,
+.row-1qtctT:hover,
+.colorwayInfoIcon,
+.colorwayCheckIcon {
+    --white-500: black !important;
+}
+
+.ColorwaySelectorBtn:hover .colorwaySelectorIcon {
+    background-color: black !important;
+}
+
+:root {
+    --mention-foreground: black !important;
+}
+/*End Accent*/`: ""}`;
+}
