+++ src/plugins/discordColorways/types.ts
@@ -1,0 +1,19 @@
+/*
+ * Vencord, a Discord client mod
+ * Copyright (c) 2023 Vendicated and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+export interface Colorway {
+    name: string,
+    import: string,
+    accent: string,
+    primary: string,
+    secondary: string,
+    tertiary: string,
+    original?: boolean,
+    author: string,
+    authorID: string,
+    colors?: string[],
+    isGradient?: boolean;
+}
